Fix correctness bugs: Market Cap scaling, scraper timeout, schema type, DB connect backoff
Batch 1 of the improvement plan:
- clean_numeric_value now scales K/M/B/T magnitude suffixes instead of
  stripping them, so Market Cap (e.g. "$34,005.3M") is stored correctly
  (~3.4e10) instead of being truncated to 34005.3. 'x' stays a plain unit.
- get_tickers() requests now use timeout=30 (matching get_stock_data), so a
  hung connection can't block the loop indefinitely.
- No Years column type is now Integer in both dtype_map and
  create_dividend_data_table (was inconsistently Float in one).
- DB connection is acquired via bounded exponential-backoff retry; a
  transient outage at connect time no longer crashes the process.

Updated tests that encoded the old (buggy) un-scaled behavior.

+++ divifilter_data_updater/db_functions.py
@@ -41,7 +41,7 @@ class MysqlConnection:
     def update_data_table_from_data_frame(self, data_table_to_update: pd.DataFrame):
         dtype_map = {
             'Symbol': String(32),
-            'No Years': Float,
+            'No Years': Integer,
             'Price': Float,
             'Div Yield': Float,
             '5Y Avg Yield': Float,
